+++ sos21-api-server/src/filter.rs
@@ -123,6 +123,10 @@ pub fn endpoints(
                     }
                 }
             },
+            / "file-sharing" {
+                / "get-file" => GET (handler::project::file_sharing::get_file),
+                / "get-file-info" => GET (handler::project::file_sharing::get_file_info),
+            },
             / "file-distribution" {
                 / "list" => GET (handler::project::file_distribution::list),
                 / "get" => GET (handler::project::file_distribution::get),
@@ -151,6 +155,10 @@ pub fn endpoints(
         },
         / "form-answer" {
             / "get" => GET (handler::form_answer::get),
+            / "file-sharing" {
+                / "get-file" => GET (handler::form_answer::file_sharing::get_file),
+                / "get-file-info" => GET (handler::form_answer::file_sharing::get_file_info),
+            }
         },
         / "user" {
             / "get" => GET (handler::user::get),
@@ -169,18 +177,8 @@ pub fn endpoints(
             / "revoke" => POST (handler::file_sharing::revoke),
             / "get-file" => GET (handler::file_sharing::get_file),
             / "get-file-info" => GET (handler::file_sharing::get_file_info),
-            / "public" {
-                / "get-file" => {noauth} GET (handler::file_sharing::public::get_file),
-                / "get-file-info" => {noauth} GET (handler::file_sharing::public::get_file_info),
-            },
-            / "project" {
-                / "get-file" => GET (handler::file_sharing::project::get_file),
-                / "get-file-info" => GET (handler::file_sharing::project::get_file_info),
-            },
-            / "form-answer" {
-                / "get-file" => GET (handler::file_sharing::form_answer::get_file),
-                / "get-file-info" => GET (handler::file_sharing::form_answer::get_file_info),
-            }
+            / "get-public-file" => {noauth} GET (handler::file_sharing::get_public_file),
+            / "get-public-file-info" => {noauth} GET (handler::file_sharing::get_public_file_info),
         },
         / "file-distribution" {
             / "create" => POST (handler::file_distribution::create),