+++ sos21-api-server/src/handler/form_answer.rs
@@ -1,2 +1,4 @@
+pub mod file_sharing;
+
 pub mod get;
 pub use get::handler as get;